// src/app/openapi/model/historico.ts
/**
 * alfa
 * beta
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: 48928310+averge@users.noreply.github.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class Historico { 
    id?: string;
    fecha?: string;
    precio?: number;
    exchangeId?: string;
   
   
}